+++ zio-http/src/test/scala/zhttp/http/middleware/CsrfSpec.scala
@@ -7,7 +7,7 @@ import zio.Ref
 import zio.test.Assertion.equalTo
 import zio.test._
 
-object CsrfMiddlewaresSpec extends DefaultRunnableSpec with HttpAppTestExtensions {
+object CsrfSpec extends DefaultRunnableSpec with HttpAppTestExtensions {
   override def spec = suite("CSRF Middlewares") {
     val app           = (Http.ok @@ csrfValidate("x-token")).getStatus
     val setCookie     = Headers.cookie(Cookie("x-token", "secret"))
